requirements.txt
absl-py==2.2.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
jwcrypto==1.5.6
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mysqlclient==2.2.7
namex==0.0.8
numpy==2.1.3
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
protobuf==5.29.4
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
